# indexer/services/comlink/public/swagger.json
{
  "components": {
...
    "schemas": {
...
      "IsoString": {
        "type": "string"
      },
...
      "HistoricalFundingResponseObject": {
        "properties": {
          "ticker": {
            "type": "string"
          },
          "rate": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "effectiveAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "effectiveAtHeight": {
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "rate",
          "price",
          "effectiveAt",
          "effectiveAtHeight"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "HistoricalFundingResponse": {
        "properties": {
          "historicalFunding": {
            "items": {
              "$ref": "#/components/schemas/HistoricalFundingResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "historicalFunding"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "TransferType": {
        "enum": [
          "TRANSFER_IN",
          "TRANSFER_OUT",
          "DEPOSIT",
          "WITHDRAWAL"
        ],
        "type": "string"
      },
      "TransferResponseObject": {
        "properties": {
          "id": {
            "type": "string"
          },
          "sender": {
            "properties": {
              "subaccountNumber": {
                "type": "number",
                "format": "double"
              },
              "address": {
                "type": "string"
              }
            },
            "required": [
              "address"
            ],
            "type": "object"
          },
          "recipient": {
            "properties": {
              "subaccountNumber": {
                "type": "number",
                "format": "double"
              },
              "address": {
                "type": "string"
              }
            },
            "required": [
              "address"
            ],
            "type": "object"
          },
          "size": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "createdAtHeight": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TransferType"
          },
          "transactionHash": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sender",
          "recipient",
          "size",
          "createdAt",
          "createdAtHeight",
          "symbol",
          "type",
          "transactionHash"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "TransferResponse": {
        "properties": {
          "transfers": {
            "items": {
              "$ref": "#/components/schemas/TransferResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "transfers"
        ],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Indexer API",
    "version": "v1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/historicalFunding/{ticker}": {
      "get": {
        "operationId": "GetHistoricalFunding",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalFundingResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "effectiveBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "effectiveBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          }
        ]
      }
    },
...
    "/transfers": {
      "get": {
        "operationId": "GetTransfers",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "https://jinx-testnet.imperator.co/v4",
      "description": "Public Testnet"
    }
  ]
}